requirements.txt
webcolors
shortuuid
stanza
spacy
pytesseract
opencv-python
pdf2image
pathlib
pandas
poppler-utils
temp
pillow
textract
matplotlib
regex
fitz
pymupdf==1.19.0
table_ocr
